Stoneの弱点の表示を追加、 hpの減少に合わせて色を変化させていくのがまだ微妙

+++ Source/PlayScene/Stage/StageSearch.cpp
@@ -183,7 +183,7 @@ void StageSearch::ReadMap(std::string fileName)
 		map.push_back(mapLine);
 	}
 	delete csv;
-	mapOffset = map.size() / 2;
+	mapOffset = (int)map.size() / 2;
 }
 
 void StageSearch::DrawSearchWay()
